feat: compute hypervolume within a cuboid
An optional parameter `max_point` is added to the hypervolume()
function.  When `max_point` is given, the return value of
hypervolume() is the hypervolume that lies within the cuboid
defined by `reference_point` (min coordinate) and `max_point`
(max coordinate).   When `max_point` is not give, the behavior
is unchanged.

To Do: remove hard_coded_dimensions